+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Added
+- Added `FlagOrValue::{is_none, is_flag, is_value, into_value, as_value}`
+
 ## [0.8.0] - 2023-09-18
 ### Changed
 - Renamed `Attribute` to `FromAttr` to not conflict with `syn::Attribute`.